request_form_slackbot/requirements.txt
boto3==1.35.6
botocore==1.35.11
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.1
httplib2==0.22.0
idna==3.8
jmespath==1.0.1
numpy==2.1.0
oauth2client==4.1.3
oauthlib==3.2.2
pandas==2.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyMySQL==1.1.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.10.2
six==1.16.0
slack-bolt==1.18.1
slack_sdk==3.31.0
tzdata==2024.1
urllib3==2.2.2


